src/hero-training/pyproject.toml
[project]
name = "hero-training"
version = "0.1.0"
description = "Ultralytics training CLI for Nexus hero detectors"
authors = [{name="Nexus Replay"}]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "ultralytics>=8.3.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.1",
    "ruff>=0.6.0"
]
cuda-cu124 = [
    "torch==2.6.0",
    "torchvision==0.21.0"
]
